ISPN-9657 Avoid split packages
* Export more specific packages
* Move counter Utils to impl package
* Move JDBC UnitTestDatabaseManager from fwk package
