+++ counter/pom.xml
@@ -65,8 +65,7 @@
                         <Export-Package>
                             ${project.groupId}.counter.EmbeddedCounterManagerFactory;version=${project.version};-split-package:=error,
                             ${project.groupId}.counter.configuration.*;version=${project.version};-split-package:=error,
-                            ${project.groupId}.counter.logging.*;version=${project.version};-split-package:=error,
-                            ${project.groupId}.counter.util.*;version=${project.version};-split-package:=merge-last
+                            ${project.groupId}.counter.logging.*;version=${project.version};-split-package:=error
                         </Export-Package>
                         <Include-Resource>
                             {maven-resources},
