+++ multimap/pom.xml
@@ -114,7 +114,8 @@
                 <configuration>
                     <instructions>
                         <Export-Package>
-                            ${project.groupId}.multimap.*;version=${project.version};-split-package:=merge-last
+                            !${project.groupId}.multimap.api.*;version=${project.version};-split-package:=error,
+                            ${project.groupId}.multimap.*;version=${project.version};-split-package:=error
                         </Export-Package>
                        <Include-Resource>
                             {maven-resources},
