+++ persistence/jdbc/src/test/java/org/infinispan/persistence/jdbc/stringbased/JdbcStringParallelIterationTest.java
@@ -3,7 +3,7 @@ package org.infinispan.persistence.jdbc.stringbased;
 import org.infinispan.configuration.cache.ConfigurationBuilder;
 import org.infinispan.persistence.ParallelIterationTest;
 import org.infinispan.persistence.jdbc.configuration.JdbcStringBasedStoreConfigurationBuilder;
-import org.infinispan.test.fwk.UnitTestDatabaseManager;
+import org.infinispan.persistence.jdbc.UnitTestDatabaseManager;
 import org.testng.annotations.Test;
 
 /**
